11-05-2022 : Mise à jour de lla timezone et mise à jour des decimal

+++ database/migrations/2013_12_18_051359_create_categories_table.php
@@ -17,7 +17,7 @@ class CreateCategoriesTable extends Migration
             $table->id();
             $table->string('code');
             $table->string('libelle');
-            $table->double('montant')->default(0);
+            $table->decimal('montant', 18, 2)->default(0);
             $table->timestamps();
         });
     }
